Edit content and styles according to UX figma
Changed the content to match what was placed into the Figma design from
the UX team. Note that the image received to be used between the rest of
the content and the grey background was incorrect. This image has a
white background which will not allow the grey color to be layered
behind and still show. For now the image is commented out until the
proper image is obtained.